requirements.txt
python-nmap
rich
textual
flask
reportlab
matplotlib
apscheduler
